requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
boto3==1.34.21
botocore==1.34.21
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0
django-import-export==3.3.7
django-storages==1.14.2
et-xmlfile==1.1.0
jmespath==1.0.1
lxml==5.3.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.5
pillow==10.2.0
PyMuPDF==1.24.11
python-dateutil==2.8.2
python-docx==1.1.2
PyYAML==6.0.1
s3transfer==0.10.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.0.7
xlrd==2.0.1
xlwt==1.3.0
